% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attach_GoogleData.R
\name{attach_GmapReportData}
\alias{attach_GmapReportData}
\title{Attach Data for Reports Extracted from Google Maps}
\usage{
attach_GmapReportData(reportSummary, sheetPath)
}
\arguments{
\item{reportSummary}{Report summary data frame to join google map info to.}

\item{sheetPath}{Absolute or relative path to google map data (.csv file)}
}
\value{
Report summary data frame with google map data attached
}
\description{
Attach Data for Reports Extracted from Google Maps
}

