% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attach_ReportGeoData.R
\name{attach_ReportGeoData}
\alias{attach_ReportGeoData}
\title{Attach geographic data to reports}
\usage{
attach_ReportGeoData(reportSummary, notes)
}
\arguments{
\item{reportSummary}{}

\item{notes}{}
}
\value{
None
}
\description{
Attach geographic data to reports
}

