% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attach_TripGeoData.R
\name{attach_TripGeoData}
\alias{attach_TripGeoData}
\title{Attach trip geographic data}
\usage{
attach_TripGeoData(tripSummary, matchLinksPath, matchTripsPath,
  linkType = "orcycle")
}
\arguments{
\item{tripSummary}{Trip summary table}

\item{matchLinksPath}{Absolute or relative path to map-matched trip links.}

\item{matchTripsPath}{Absolute or relative path to map-matched trips.}

\item{linkType}{Parameter used to determine if matched links are from ORcycle trips or map-matched directions from a Directions API (Google or MapBox)}
}
\value{
Trip table with geographic data attached.
}
\description{
Use attributed network to add geographic data to trip summary table.
}

