% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attach_TripTimeData.R
\name{attach_TripTimeData}
\alias{attach_TripTimeData}
\title{Attach trip time data}
\usage{
attach_TripTimeData(tripTable, trips)
}
\arguments{
\item{tripTable}{Trip summary table}

\item{trips}{'trip' table from SQL database}
}
\value{
Trip table with temporal data attached.
}
\description{
Attach temporal data to trip summary table
}

