% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attach_TripWeatherData.R
\name{attach_TripWeatherData}
\alias{attach_TripWeatherData}
\title{Attach trip weather data}
\usage{
attach_TripWeatherData(tripTable, trips, coords)
}
\arguments{
\item{tripTable}{Trip summary table}

\item{trips}{'trip' table from SQL database}

\item{coords}{'coord' table from SQL database}
}
\value{
Trip table with weather data attached.
}
\description{
Attach weather data to trip summary table
}

