% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subDataByGeoBound.R
\name{classifyCoordsGeoBound}
\alias{classifyCoordsGeoBound}
\title{Classify coordinate data by geographic boundary}
\usage{
classifyCoordsGeoBound(coords, boundShape, labelField)
}
\arguments{
\item{boundShape}{Geographc subsetting boundary file (SpatialPolygons object)}

\item{labelField}{Field name of categorization label}

\item{con}{Database connection object}
}
\value{
None
}
\description{
Classify coordinate data by geographic boundary
}

