% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossTabPlot.R
\name{crossTabPlot}
\alias{crossTabPlot}
\title{Cross tablutation plot}
\usage{
crossTabPlot(data, xVar, yVar, xlab, ylab, colorPal, xCats = levels(x),
  yCats = levels(y), px = 1280, py = 560)
}
\arguments{
\item{data}{Trip, user, or report table for plotting}

\item{xVar}{Variable divide into seperate bars}

\item{yVar}{Variable to fill sections of bars by}

\item{xlab}{Label for x variable}

\item{ylab}{Label of y variable}

\item{colorPal}{Color pallette to use for plotting}

\item{xCats}{X category names, default is levels of x variable}

\item{yCats}{Y category names, default is levels of y variable}

\item{px}{Width of plot in pixels, default is 1280 pixels}

\item{py}{Height of plot in pixels, default is 560 pixels}
}
\value{
None
}
\description{
Cross tablutation plot
}

