% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/photoSync.R
\name{downloadReportPhotos}
\alias{downloadReportPhotos}
\title{Download Report Photos}
\usage{
downloadReportPhotos(remotePath, localPath)
}
\arguments{
\item{remotePath}{The remote path from which to download photos. This requires a server connection.}

\item{localPath}{The local path to sync the remote photo repository with.}
}
\value{
None
}
\description{
Sync report photos from remote stash folder to local photo repository in the `orcycler` environment.
}

