% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpsTrimmer.R
\name{gpsTrimmer}
\alias{gpsTrimmer}
\title{GPS Trimmer}
\usage{
gpsTrimmer(rawCoords, trimDist)
}
\arguments{
\item{rawCoords}{Raw coordinate table in same format as 'coord' table in SQL database.}

\item{trimDist}{Distance to trim from origin and destination of trip.}
}
\value{
Trimmed coordinate table
}
\description{
Trims GPS traces for ORcycle trips from both beginning and end of trip by a given distance (in miles)
}
\examples{
gpsTrimmer(coords,0.05)
}

