% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputeResponses.R
\name{imputeResponses}
\alias{imputeResponses}
\title{Impute Responses}
\usage{
imputeResponses(data, impCols)
}
\arguments{
\item{data}{Summary table to impute missing responses in.}

\item{impCols}{Columns in summary table to use in imputation, including both columns with/without missing responses. Values used for imputation must be numeric or factor.}
}
\value{
Summary table with imputed responses.
}
\description{
Wrapper for \code{\link[missForest]{missForest}} function used to impute missing responses in user, trip, and report data.
}

