% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_barPlots.R
\name{report_barPlots}
\alias{report_barPlots}
\title{Plot distribution plots for report variables}
\usage{
report_barPlots(reportSummary, reportType)
}
\arguments{
\item{reportSummary}{Prepared report data summary}

\item{reportType}{Type of report to produce plots for (options are 0 or 1)}
}
\value{
None
}
\description{
Plot distribution plots for report variables
}

