% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportAnalysis.R
\name{reportAnalysis}
\alias{reportAnalysis}
\title{Report Analysis}
\usage{
reportAnalysis(db = "test", db_path = "data/db_credentials.json")
}
\arguments{
\item{db}{String indicating which db to use, default is "test", also can be "release"}

\item{db_path}{Relative or absolute file path to db_credentials.json}
}
\value{
None
}
\description{
Initializes database connection and calls series of functions used to analyze ORcycle trip data.
}

