% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportPrep.R
\name{reportPrep}
\alias{reportPrep}
\title{Prepare report data}
\usage{
reportPrep(reportType, users, notes, noteResponses, questions, answers, qaMap)
}
\arguments{
\item{reportType}{Type of report to prepare summary for (0 or 1)}

\item{users}{User data in format of 'user' table from SQL database}

\item{notes}{Note data in format of 'note' table from SQL database}

\item{noteResponses}{Note response data in format of 'noteResponse' table from SQL database}

\item{questions}{Question cross reference table from SQL database}

\item{answers}{Answer cross reference table from SQL database}

\item{qaMap}{Question-answer map cross reference table from SQL database}
}
\value{
None
}
\description{
Prepare report data
}

