% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleMapUtils.R
\name{route2Shape}
\alias{route2Shape}
\title{Convert decoded Google Directions polyline into SpatialLines object}
\usage{
route2Shape(directions)
}
\arguments{
\item{directions}{Decoded polyline directions from Google Directions API}
}
\value{
SpatialLines object representing route received from Google Directions API
}
\description{
Convert decoded Google Directions polyline into SpatialLines object
}

