% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveOutReports.R
\name{saveReportsToExcel}
\alias{saveReportsToExcel}
\title{Save Report Summary to Excel}
\usage{
saveReportsToExcel(crashSummary, issueSummary, notes, saveFolder)
}
\arguments{
\item{crashSummary}{Crash report summary table}

\item{issueSummary}{Issue report summary table}

\item{notes}{Note table from SQL database}

\item{saveFolder}{Relative or Absolute path to output folder}
}
\value{
None
}
\description{
Save Report Summary to Excel
}

