% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarityWeight.R
\name{similarityWeight}
\alias{similarityWeight}
\title{Weight trips based on link similarity}
\usage{
similarityWeight(matchLinksPath, trips, tripMatchSummmary)
}
\arguments{
\item{matchLinksPath}{Absolute or relative path to map matched links}

\item{trips}{'trip' table in format of SQL table}

\item{tripMatchSummmary}{Summary table of matched trips}
}
\value{
Similarity weights
}
\description{
Weight trips based on link similarity
}

