% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subDataByGeoBound.R
\name{subDataByGeoBound}
\alias{subDataByGeoBound}
\title{Subset data by geographic boundary}
\usage{
subDataByGeoBound(con, boundShape, folderName, startDate, endDate,
  privacy = FALSE, include_qa = TRUE)
}
\arguments{
\item{con}{Database connection object}

\item{boundShape}{Geographc subsetting boundary (SpatialPolygons object)}

\item{folderName}{Name of folder to output}

\item{startDate}{Start date of subset in YYYY-MM-DD format}

\item{endDate}{End date of subset in YYYY-MM-DD format}

\item{privacy}{Boolean indicating whether to include user IDs (not included if TRUE)}

\item{include_qa}{Boolean indicating if question answer dictionaries will be included in output folder}
}
\value{
None
}
\description{
Subset data by geographic boundary
}

