% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/photoSync.R
\name{syncPhotos}
\alias{syncPhotos}
\title{Synchronize Photos in Local Repository with Dropbox Folder}
\usage{
syncPhotos(localPath, dropPath)
}
\arguments{
\item{localPath}{Absolute or Relative Path to local repository}

\item{dropPath}{Path to dropbox photo repository relative to dropbox root folder}
}
\value{
None
}
\description{
This will synchronize the photos between a local repository folder in the `orcycler` environment and the dropbox folder where report photos will be hosted.
}

