% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trafficStress.R
\name{trafficStress}
\alias{trafficStress}
\title{Level of Traffic Stress calculation}
\usage{
trafficStress(segReports)
}
\arguments{
\item{segReports}{Reports made on segments of transportation facility (i.e. not at intersections)}
}
\value{
Segment report data frame with Level Of Traffic Stress (LTS) calculated
}
\description{
Level of Traffic Stress calculation
}
\references{
The logical and mathematical calculations in this function were based upon the "Low Stress Bicycling and Network Connectivity" report developed by Mineta Transportation Institute http://transweb.sjsu.edu/PDFs/research/1005-low-stress-bicycling-network-connectivity.pdf and the Analysis Procedures Manual developed by ODOT: http://www.oregon.gov/ODOT/TD/TP/APM/APMv2.pdf
}

