% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tripAnalysis.R
\name{tripAnalysis}
\alias{tripAnalysis}
\title{Trip Analysis}
\usage{
tripAnalysis(db = "test", db_path = "data/db_credentials.json")
}
\arguments{
\item{db}{String indicating which db to use, default is "test", also can be "release"}

\item{db_path}{Relative or absolute file path to db_credentials.json}
}
\value{
Trip analysis results
}
\description{
Initializes database connection and calls series of functions used to analyze ORcycle trip data.
}

