% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tripPrep.R
\name{tripPrep}
\alias{tripPrep}
\title{Prepare trip data table}
\usage{
tripPrep(trips, tripResponses, answers, questions, qaMap, coords)
}
\arguments{
\item{trips}{'trip' table from SQL database}

\item{tripResponses}{'tripResponse' table from SQL database}

\item{answers}{'answer' table from SQL database}

\item{questions}{'question' table from SQL database}

\item{qaMap}{'question_answer_relate' table from SQL database}

\item{coords}{'coord' table from SQL database}
}
\value{
Flat trip table for further exploration/analysis.
}
\description{
Proccesses all trip-related SQL data into a single flat table for further exploratory and analysis tasks.
}

