% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateMapMatchCoords.R
\name{updateMapMatchCoords}
\alias{updateMapMatchCoords}
\title{Update coordinate file for map matching}
\usage{
updateMapMatchCoords(coords, matchLinksPath)
}
\arguments{
\item{coords}{'coord' table from SQL database}

\item{matchLinksPath}{Relative or absolute path to existing matched link set.}
}
\value{
None
}
\description{
This function updates queued coordinates for map matching. Existing map match links are considered to filter coordinate file to only trips that have not been matched.
}

