% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateMapMatchLinks.R
\name{updateMapMatchLinks}
\alias{updateMapMatchLinks}
\title{Update links and trips files from map matching.}
\usage{
updateMapMatchLinks(matchLinksPath, matchTripsPath)
}
\arguments{
\item{matchLinksPath}{Relative or absolute path to existing matched link set.}

\item{matchTripsPath}{Relative or absolute path to existing matched trip set.}
}
\value{
None
}
\description{
This function updates map-matched links and trips files with results from python scripts stored in local PostgreSQL database.
}

