% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/user_barPlots.R
\name{user_barPlots}
\alias{user_barPlots}
\title{User response plotting}
\usage{
user_barPlots(userSummary, outDir)
}
\arguments{
\item{userSummary}{User summary table outputted from \code{\link{userPrep}}}

\item{outDir}{Relative or absolute directory destination for resultant plots}
}
\value{
None
}
\description{
Plots all user response distributions to 'results' folder
}

