% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/userAnalysis.R
\name{userAnalysis}
\alias{userAnalysis}
\title{User Analysis}
\usage{
userAnalysis(db_info)
}
\arguments{
\item{db_info}{Absolute or relative path to db_info.txt file.}
}
\value{
User analysis results
}
\description{
Initializes database connection and calls series of functions used to analyze ORcycle user data.
}

