% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/userPrep.R
\name{userPrep}
\alias{userPrep}
\title{Prepare user data table}
\usage{
userPrep(users, userResponses, answers, questions, qaMap, trips, notes)
}
\arguments{
\item{users}{'user' table from SQL database}

\item{userResponses}{'userResponse' table from SQL database}

\item{answers}{'answer' table from SQL database}

\item{questions}{'question' table from SQL database}

\item{qaMap}{'question_answer_relate' table from SQL database}

\item{trips}{'trip' table from SQL database}

\item{notes}{'note' table from SQL database}
}
\value{
Flat user table for further exploration/analysis.
}
\description{
Proccesses all user-related SQL data into a single flat table for further exploratory and analysis tasks.
}

