% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap_sentance.R
\name{wrap_sentence}
\alias{wrap_sentence}
\title{Wrap long strings}
\usage{
wrap_sentence(stringVector, width)
}
\arguments{
\item{stringVector}{Vector of strings to split onto multiple lines}

\item{width}{Line width to trim strings to (in number of characters)}
}
\value{
Returns strings split onto multiple lines if they exceed the width parameter.
}
\description{
Wraps long strings onto multiple lines for plotting
}
\author{
RickyB
}
\references{
\url{http://stackoverflow.com/questions/7367138/text-wrap-for-plot-titles}
}

