% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trip_barPlots.R
\name{trip_barPlots}
\alias{trip_barPlots}
\title{Trip response plotting}
\usage{
trip_barPlots(tripSummary, outDir)
}
\arguments{
\item{tripSummary}{Trip summary table outputted from \code{\link{tripPrep}}}

\item{outDir}{Relative or absolute directory destination for resultant plots}
}
\value{
None
}
\description{
Plots all trip response distributions to 'results' folder
}

